+++ KFXAdditions/Classes/NSDate+KFXAdditions.m
@@ -9,28 +9,23 @@
 //--------------------------------------------------------
 #pragma mark - Comparison Convience methods
 //--------------------------------------------------------
--(BOOL)kfx_isLaterThanDate:(NSDate *)anotherDate{
-    
-    NSComparisonResult result = [self compare:anotherDate];
-    if (result == NSOrderedAscending) {
-        return NO;
-    }else if (result == NSOrderedSame){
-        return NO;
-    }else{
-        return YES;
-    }
-}
-
 -(BOOL)kfx_isEarlierThanDate:(NSDate *)anotherDate{
     
-    NSComparisonResult result = [self compare:anotherDate];
-    if (result == NSOrderedDescending) {
-        return NO;
-    }else if (result == NSOrderedSame){
-        return NO;
-    }else{
-        return YES;
-    }
+	return [self compare:anotherDate] == NSOrderedAscending;
+}
+
+-(BOOL)kfx_isLaterThanDate:(NSDate *)anotherDate{
+	
+	return [self compare:anotherDate] == NSOrderedDescending;
+}
+
+-(BOOL)kfx_isEarlierThanOrEqualToDate:(NSDate*)anotherDate{
+	
+	return [self compare:anotherDate] != NSOrderedDescending;
+}
+
+-(BOOL)kfx_isLaterThanOrEqualToDate:(NSDate*)anotherDate{
+	return [self compare:anotherDate] != NSOrderedAscending;
 }
 
 -(BOOL)kfx_isBetweenStartDate:(NSDate *)startDate andEndDate:(NSDate *)endDate{
@@ -78,6 +73,15 @@
 			&& [compA minute]==[compB minute]
 			);
 }
+
+-(BOOL)kfx_isTimeEqualToDate:(NSDate *)anotherDate{
+	
+	NSDateComponents *compA = [self kfx_currentCalendarDateComponents];
+	NSDateComponents *compB = [anotherDate kfx_currentCalendarDateComponents];
+	return ([compA hour]==[compB hour]
+			&& [compA minute]==[compB minute]);
+}
+
 
 
 //--------------------------------------------------------