+++ KFXAdditions/Classes/UIView+KFXAdditions.h
@@ -6,6 +6,12 @@
 @interface UIView (KFXAdditions)
 
 //--------------------------------------------------------
+#pragma mark Init
+//--------------------------------------------------------
+/// Assumes the nib has the same name as the class. Loads nib with nil owner and options
++(instancetype)kfx_instantiateFromNib;
+
+//--------------------------------------------------------
 #pragma mark - First Responder
 //--------------------------------------------------------
 -(UIView*)kfx_findFirstResponder;