+++ KFXAdditions/Classes/UIView+KFXAdditions.m
@@ -6,6 +6,16 @@
 @implementation UIView (KFXAdditions)
 
 //--------------------------------------------------------
+#pragma mark Init
+//--------------------------------------------------------
++(instancetype)kfx_instantiateFromNib{
+	return [[NSBundle mainBundle]loadNibNamed:NSStringFromClass([self class])
+										owner:nil
+									  options:nil].firstObject;
+}
+
+
+//--------------------------------------------------------
 #pragma mark - First Responder
 //--------------------------------------------------------
 -(UIView *)kfx_findFirstResponder{